+++ source/engine/ui/editor/impl/EditorUI.h
@@ -1,11 +1,10 @@
 #pragma once
 
-#include "ncengine/ecs/Ecs.h"
-#include "windows/CreateEntityDialog.h"
 #include "windows/FpsOverlay.h"
 #include "windows/Inspector.h"
-#include "windows/ModalDialog.h"
 #include "windows/SceneGraph.h"
+#include "windows/dialogs/CreateEntityDialog.h"
+#include "windows/dialogs/SceneDialogs.h"
 
 #include "ncengine/ecs/Ecs.h"
 #include "ncengine/module/ModuleProvider.h"
