+++ source/engine/ui/editor/impl/windows/CMakeLists.txt
@@ -2,7 +2,6 @@ target_sources(${NC_ENGINE_LIB}
     PRIVATE
         EntityContextMenu.cpp
         Inspector.cpp
-        ModalDialog.cpp
         SceneGraph.cpp
 )
