+++ source/engine/ui/editor/impl/windows/dialogs/CMakeLists.txt
@@ -1,4 +1,5 @@
 target_sources(${NC_ENGINE_LIB}
     PRIVATE
+        CreateEntityDialog.cpp
         SceneDialogs.cpp
 )
